requirements.txt
mysql-connector-python==8.1.0
python-dotenv==1.0.0
xmlschema==2.4.0